ELECTRON-32 - Added period attribute
