+++ js/preload/preloadMain.js
@@ -126,17 +126,20 @@ function createAPI() {
         /**
          * allows JS to register a activity detector that can be used by electron main process.
          * @param  {Object} activityDetection - function that can be called accepting
+         * @param  {Object} period - minimum user idle time in millisecond
          * object: {
+         *  period: Number
          *  systemIdleTime: Number
          *  }
          */
-        registerActivityDetection: function(activityDetection) {
+        registerActivityDetection: function(period, activityDetection) {
             if (typeof activityDetection === 'function') {
                 local.activityDetection = activityDetection;
 
                 // only main window can register
                 local.ipcRenderer.send(apiName, {
-                    cmd: apiCmds.registerActivityDetection
+                    cmd: apiCmds.registerActivityDetection,
+                    period: period
                 });
             }
         },
@@ -177,7 +180,7 @@ function createAPI() {
     // listen for user activity from main process
     local.ipcRenderer.on('activity', (event, arg) => {
         if (local.activityDetection && arg && arg.systemIdleTime) {
-            local.activityDetection(arg.systemIdleTime, arg.isUserActive);
+            local.activityDetection(arg.systemIdleTime);
         }
     });
